+++ src/Oro/Bundle/FrontendBundle/Tests/Unit/DependencyInjection/Compiler/FrontendApiPassTest.php
@@ -21,7 +21,6 @@ class FrontendApiPassTest extends \PHPUnit\Framework\TestCase
     private const PROCESSORS = [
         'oro_api.collect_resources.load_dictionaries',
         'oro_api.collect_resources.load_custom_entities',
-        'oro_api.collect_resources.add_excluded_actions_for_dictionaries',
         'oro_api.options.rest.set_cache_control',
         'oro_api.rest.cors.set_allow_origin',
         'oro_api.rest.cors.set_allow_and_expose_headers',